Changed datatype.

+++ Section7/BankAccount.java
@@ -19,7 +19,7 @@ public class BankAccount {
         BankAccount.deposit = deposit;
     }
     //Creating a constructor to represent the act of checking your balance.
-    public double balance()
+    public int balance()
     {
         //Returning the current balance.
         return balance;
